fix(zk,height,pedersen,harness): MOC close-out items 1-8
Item 1 (HIGH): Zip-truncation gap — proofs.len()==zkp_table.len() guard
  + per-call length guard in verify_core_tx_with_tables; 0x06 added to
  verify_single_tx mempool proof-requiring selectors. Test:
  test_proof_metadata_length_mismatch_rejected.

Item 2 (MEDIUM): expected_reward(u64) — dropped as u32 casts across 12
  call sites in 10 files (block.rs, chain_state.rs, block_acceptor.rs,
  lib.rs, miner.rs, stratum.rs, blockchain.rs, mm_rpc.rs, model.rs,
  entrypoint/mod.rs, pow_reward_v1.rs, heavyweight_pipeline.rs).

Item 3 (LOW): Dead tx_binding witness removed from mint_v1.zk and both
  prover sites in transfer_v1/proof.rs. Witness was shadowed by in-circuit
  computation at line 100.

Item 4 (MEDIUM): pedersen_commitment_base deprecated (uses NullifierK
  generator incompatible with all zkas circuits); removed from Python
  bindings exports and crypto/mod.rs re-export.

Item 5 (MEDIUM): Heavyweight harness exec() switched from
  apply_block_with_uncles (WASM-bypass, Defect-3 rejected for height>1)
  to production accept_block path.

Item 6 (LOW): Duplicate nullifier in FeeCollectParamsV1 — ACCEPTED
  (matches PoWRewardV1 pattern, no change).

Item 7 (LOW): ZkPublicInputs comment 7->9 elements.

Item 8 (LOW): select_for_block doc fix in consensus-coinbase.md.

Build: wasm + dwow_chain + dwowd + dwow_wallet clean. OsRng zero.

Items 9 (pipeline determinism test) and 10 (mempool hardening) remain
deferred to their own plan.

+++ src/contract/native_token/src/client/pow_reward_v1.rs
@@ -253,7 +253,7 @@ impl PoWRewardCallBuilder {
 
     /// Build the PoWReward call with the standard block reward plus fees.
     pub fn build(&self) -> Result<PoWRewardCallDebris> {
-        let reward = expected_reward(self.block_height) + self.fees;
+        let reward = expected_reward(self.block_height as u64) + self.fees;
         self._build(reward)
     }
